[fields] Clean the useField hook (part 1)
