+++ packages/x-date-pickers-pro/src/SingleInputTimeRangeField/useSingleInputTimeRangeField.ts
@@ -1,9 +1,5 @@
 'use client';
-import {
-  useField,
-  useFieldInternalPropsWithDefaults,
-  PickerRangeValue,
-} from '@mui/x-date-pickers/internals';
+import { useField } from '@mui/x-date-pickers/internals';
 import { useSplitFieldProps } from '@mui/x-date-pickers/hooks';
 import { UseSingleInputTimeRangeFieldProps } from './SingleInputTimeRangeField.types';
 import { useTimeRangeManager } from '../managers';
@@ -16,23 +12,5 @@ export const useSingleInputTimeRangeField = <
 ) => {
   const manager = useTimeRangeManager(props);
   const { forwardedProps, internalProps } = useSplitFieldProps(props, 'time');
-  const internalPropsWithDefaults = useFieldInternalPropsWithDefaults({
-    manager,
-    internalProps,
-  });
-
-  return useField<
-    PickerRangeValue,
-    TEnableAccessibleFieldDOMStructure,
-    typeof forwardedProps,
-    typeof internalPropsWithDefaults
-  >({
-    forwardedProps,
-    internalProps: internalPropsWithDefaults,
-    valueManager: manager.internal_valueManager,
-    fieldValueManager: manager.internal_fieldValueManager,
-    validator: manager.validator,
-    valueType: manager.valueType,
-    getOpenPickerButtonAriaLabel: manager.internal_useOpenPickerButtonAriaLabel(),
-  });
+  return useField({ manager, forwardedProps, internalProps });
 };
