+++ packages/x-date-pickers/src/DateTimeField/useDateTimeField.ts
@@ -1,9 +1,8 @@
 'use client';
-import { useField, useFieldInternalPropsWithDefaults } from '../internals/hooks/useField';
+import { useField } from '../internals/hooks/useField';
 import { UseDateTimeFieldProps } from './DateTimeField.types';
 import { useSplitFieldProps } from '../hooks';
 import { useDateTimeManager } from '../managers';
-import { PickerValue } from '../internals/models';
 
 export const useDateTimeField = <
   TEnableAccessibleFieldDOMStructure extends boolean,
@@ -13,23 +12,5 @@ export const useDateTimeField = <
 ) => {
   const manager = useDateTimeManager(props);
   const { forwardedProps, internalProps } = useSplitFieldProps(props, 'date-time');
-  const internalPropsWithDefaults = useFieldInternalPropsWithDefaults({
-    manager,
-    internalProps,
-  });
-
-  return useField<
-    PickerValue,
-    TEnableAccessibleFieldDOMStructure,
-    typeof forwardedProps,
-    typeof internalPropsWithDefaults
-  >({
-    forwardedProps,
-    internalProps: internalPropsWithDefaults,
-    valueManager: manager.internal_valueManager,
-    fieldValueManager: manager.internal_fieldValueManager,
-    validator: manager.validator,
-    valueType: manager.valueType,
-    getOpenPickerButtonAriaLabel: manager.internal_useOpenPickerButtonAriaLabel(),
-  });
+  return useField({ manager, forwardedProps, internalProps });
 };
