+++ packages/x-date-pickers/src/TimeField/useTimeField.ts
@@ -1,9 +1,8 @@
 'use client';
-import { useField, useFieldInternalPropsWithDefaults } from '../internals/hooks/useField';
+import { useField } from '../internals/hooks/useField';
 import { UseTimeFieldProps } from './TimeField.types';
 import { useSplitFieldProps } from '../hooks';
 import { useTimeManager } from '../managers';
-import { PickerValue } from '../internals/models';
 
 export const useTimeField = <
   TEnableAccessibleFieldDOMStructure extends boolean,
@@ -13,23 +12,5 @@ export const useTimeField = <
 ) => {
   const manager = useTimeManager(props);
   const { forwardedProps, internalProps } = useSplitFieldProps(props, 'time');
-  const internalPropsWithDefaults = useFieldInternalPropsWithDefaults({
-    manager,
-    internalProps,
-  });
-
-  return useField<
-    PickerValue,
-    TEnableAccessibleFieldDOMStructure,
-    typeof forwardedProps,
-    typeof internalPropsWithDefaults
-  >({
-    forwardedProps,
-    internalProps: internalPropsWithDefaults,
-    valueManager: manager.internal_valueManager,
-    fieldValueManager: manager.internal_fieldValueManager,
-    validator: manager.validator,
-    valueType: manager.valueType,
-    getOpenPickerButtonAriaLabel: manager.internal_useOpenPickerButtonAriaLabel(),
-  });
+  return useField({ manager, forwardedProps, internalProps });
 };
