+++ packages/x-date-pickers/src/internals/hooks/useField/index.ts
@@ -2,7 +2,7 @@ export { useField } from './useField';
 export type {
   FieldValueManager,
   UseFieldInternalProps,
-  UseFieldParams,
+  UseFieldParameters as UseFieldParams,
   UseFieldResponse,
   FieldChangeHandler,
   FieldChangeHandlerContext,
